init-my-stuff: fix archive ownership problems
When this file is executed on a remote host but from the localhost, the
ownership of the files remains with the localhost user. Discovered that
the "--owner=$username" option is for creating the archive, not for
extracting it. The option to use when extracting the archive is
"--no-same-user". Commented out a few once active lines to keep around
for history's sake.