createoptions: initial entry
Utility for adding bash code to parse command line options to a bash
script. There's a lot of repetition in creating options parsing code,
and there's no easy way I know of to make a template that can expand
dynamically, so a code generator seemd the best way to go.