renpat: add logic to numbering to pad with zeros
When receiving patch sets with PATCH xx/yyy, the number of digits
can present a problem when applying patches, because they will be
applied out of order. For example, patch 01/101 gets applied, then
patch 101/101 gets applied, then patch 02/101, etc.

This is commit just lays the groundwork, but does not complete
the fix.