patchreview: exit if git checkout returns nonzero
Previous behavior was to force the checkout when user selected a
different branch, which would discard any changes the user had
not committed. This gives the user an opportunity to either stash
or commit the changes before checking out a different branch.

Also simplified the clean_patches function, but left the old code
as a comment for historical reference or possible future use.

In menu_parser(), only count the patch files in the $indir. No need
to count files that are not patches.