patchreview - create a session environment
Every time a patch set is applied to the current branch, update the
session file with the 'git log --online' of the last git am commit.

If the user wants to restore one of the heads, the list of saved heads
will be presented, from which the user can select.