patcmp: cleanup
Don't want to echo the return value for global variable index, but
rather set its value in the function.

Danger is when a variable is waiting for its new value via an echo
or print statement from a function, and the function is trying to
access buffered I/O. The script will just hang.