bootloop: Allow user options
Changed the comments into output on the logged-in terminal. The
interaction between the script and the user takes place on the
logged-in terminal. After the bootloop starts, reports are sent
to /dev/ttySx serial console.

User has the options to exit the script and to change the timeout
between reboots.