distro2iso: added parse_config()
Want to parse the .treeinfo config file of the DVD to be able to create
names without having to use so many arguments to distro2iso. The user
probably wants the custom DVD to have the same name as the original,
anyway. One argument will be used for a user-supplied variant name,
perhaps as an option.