patbatcmpmgr: help fixes and added option for outfile
Put the main program inside a loop so the user can experiment to get
the output s/he wants before exiting.

If this script is called by another script, and the caller passes
-o <file>, then use that for the output file.
Else use the ofile saved in the project file.
If that's an empty string, use the default /dev/stdout

Changed the r and u options to upper case, R  and U.

Expanded the help text, so piping to less for easier reading.