patbatcmp: Add verbose level to output file
Adding verbose level to the output file gives a calling routine
a way to understand the contents of the file.

Also:
Fixed verbose and vverbose logic. If vverbose is true, then verbose
must be true. Conversely, if verbose is false, then vverbose must
also be false.

Added a boolean to indicate whether output is going to /dev/stdout.

Changed the format of the output so that the line naming the files
being compared always starts with "PATCH:" to make parsing by any
callers easier.