Notify-wait for git describe --tags --abbrev=0
I noticed this was taking some time, and it might lead the user to
think the script is wedged. The additional code spawns the "git
describe" using '&' and prints a dot every second while waiting for
"git describe" to write its output into shared memory at /dev/shm
When we see /dev/shm has the file in it, we can continue execution.

	modified:   lib/patch-common.source
	modified:   patbatcmpmgr
	modified:   patchdevelop
	modified:   patchreview
	modified:   patcmp