Provide the user a choice of editors.
Allow the user to select the editor of choice from a list. Currently
there are only Emacs and Vim supported.

There were some ancillary changes made for confomrity. get_mailclient()
was removed, as it was unused. Instead, prj_get_item() was called
directly.

There was a bug in patcmp when returning from patbatcmpmgr. The
mismatch_file needed to be set. Previous to this patch, it was only set
when passed as an option.

	modified:   lib/patch-common.source
	modified:   patchdevelop
	modified:   patchreview
	modified:   patcmp
	modified:   renmutpat

Signed-off-by: Dave Arcari <dacari@redhat.com>
Signed-off-by: Tony Camuso <tcamuso@redhat.com>