lib/ui.source: refactor getyn, ui_inarray, remove cruft from getuser
getyn() has needed attention for a long time. Removed all the stuff
intended for exiting on 'q' and spawning a new shell on 'x'. Decided
it was best left to the implementer to provide those options.
Also used double underscores in the local variables to avoid what
seems to be bash stepping on itself from time to time. Might be
a bash bug.

getuser(), removed the 'q' and 'x' options.

ui_inarray, added an optional output argument to take the index of
the item, if it was found in the array, -1 if it was not. Provided
new exammples