patcmp: refactor, fix bugs, and add functionality
This checkin should provide complete integration with batch compare.

Rewrote backup(), renamed getpatchnumber to get_patch_number and
rewrote it to integrate batch compare mismatch lists.

Added backup_wrapper() to package the globals up and send them to
backup and to also make the necessary adjustments when operating
in mismatch mode.

Added notify_mismatch_mode() to keep the user's selection of specific
patch numbers to within the list of mismatched patches, but have not
gotten this part functioning yet. The get_patch_number() function
works well without it, and I may not even bother to get it working.

Refactored replace_commit() to use the new backup() rather than
trying to backup on its own.

Abstracted the editor commands so that I would not need to have the
editor select case statement inside the while loop.

Changed the way the range checker reports out of range patch numbers
to work with mismatch mode. Also put it at the top of the loop so
we could avoid trying to print filenames that don't exist.

A bit of refactoring of the main menu and considerable refactoring
of the menu's case statement to adjust for the mismatch adaptations
and refactoring elsewhere.