Add support for switchable hash size and tri-state comparison mode
Comparison mode was two semi-mutually-exclusive booleans, so this
refactor changes the comparison mode to a tri-state object The three
states are Batch, Interactive Only, and Off. The user cycles through
the states by pressing the associated key.

Hash size was fixed at {7,40}, so that any 7-digit number was
considered. This was a hassle when the patch submitter was diligent
to provide 40-digit hash numbers and there were other 7-or-more-digit
numbers in the commit log.

This sets the default to 40-digit commit hashes, but is switchable to
{7,40} (minimum 7, maximum 40) for those times when you encounter
7 or 8 digit commit hashes.

modified:   lib/gitutilities.source
modified:   lib/patch-common.source
modified:   lib/ui.source
modified:   patchdevelop
modified:   patchreview