patbatcmpmgr: refactoring and bug fixes
Changed init_defaults to init_booleans. The non-boolean values will
be initialized elsewhere.

Major bug in init_booleans was that the wrong variables were used
as indexes into the project file.

Moved the initialization of the ofile to before entering the main
loop.

Moved the control_c trap to the top of the file.