patchreview: added Nuke option by request
Also changed the G option to reset the git head to the most recent
tag of the current branch.

The H option now does what the G option used to do, that is, present
a selection of recent git heads for the user to choose.

Deprecated the P and W options, as they are almost never used.