lib/ui.source:ui_getnewdir() day-1 bug
If user entered a directory that didn't exist, and didn't want to
create it, the function was stuck in an infinite loop asking for
a directory, even if you gave it a valid one that exists.

Problem was waiting on a variable that was no longer used, having
been commented out by earlier changes.

All the more reason not to leave old code lying around as "historical
commentary".