kabiscan: updates and improvements
Wanted to make this installable by rpm, so the directories had to be
relative, rather than absolute. Now this utility and its ancillary
scripts can be located anywhere in the executable path and function
correctly, since the paths are now relative to where this script is
installed.

Decided that the correct control-c behavior is always to exit. Trapping
control-c to re-enter the program might be disconcerting for most users.
Typing control-c now does a save of the current parameters, removal of
temp files, and an exit with the control-c code defined in ui.source.