lib/kabiscan.source: added help script
Created a script as a help for kabiscan.
Changed the rotation, putting Symbols first and Exported functions
second.
Added the third parameter to the prj_init_prjfile in order to bypass
the blank lines check, so that the other values would remain sticky
even when there were no strings in the "name" or "mask" fields.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>