patchreview/devleop: make automatic patch compare the default
It was annoying having to run patcmpmanager all the time. This runs it
without user intervention as the default, and adds the option to invoke
the patbatcmpmgr menu to setup user preferences.

Also updated the man pages for patchreview and patchdevelop.

lib/patch-common.source
man/patchdevelop.txt
man/patchreview.txt
patbatcmpmgr
patchdevelop
patchreview