patchreview/develop: add tkdiff as a diff editor choice
Refactored the whole editor select process when adding this in.
More streamlined, more maintainable, and more options.

Changed the Editor line in patchreview and patchdevelop to say
"Diff Editor"

Changed loop_range() in ui.source to accept single char for numbers
less than 10.

Added tkdiff to patcmp. Also changed the names of the editors to
lower case, and vim to vimdiff.