patchreview/develop: make patch -p1 an apply option
patch -p1 was a failover option. Making it an apply option cleans up the
interface a bit and is more logically consistent. The action on patch
failure will now always be to exit the script without doing violence or
damage to the tree.

Made the names of the patch (RHEL) and work (upstream) directories
variables, so they can be set in just one place. Replaced occurences of
these directory name strings with their variable names instead.

Added a new Environment statistic, the current head of the current
branch. Now you are given the last tag, the current branch and the
current head of the current branch. This should make it a lot easier to
determine where you are in the patch review/develop process.

The current head is truncated to 79 characters, so it fits in a small
terminal window.

Bumped the version to 6.0