patchrevdev: problem with patch/kernel matching code
Refactored check_patchvbranch() and show_branch_mismatch() to simplify
and make more maintainable as a side effect of fixing the bug.

But was that the lasttag global variable was not updated after calling
git_setbranch().

Also moved the init of the tag variable in check_patchvbranch() to
inside the while loop in order to refresh it on each pass.