patchrevdev: refactored lib/patch_common::check_patchvbranch()
This routine is supposed to check the distro of the patches against the
kernel of the current git branch. If they mismatch, the user is
presented with the list of current branches available and asked if s/he
would like to change to the appropriate branch.

Two problems.

1. There was no re-check after the user selected a different branch.
   This was fixed with a recursive all to check_patchvbranch to assure
   that the correct branch was selected.

2. RHEL and Pegas were not always the first words in the patch filename.
   This was fixed by tokenizing the patch filename and looking for the
   RHEL or Pegas word. Additionally, some people leave spaces or put
   hyphens between the distro and the release number. Code was added to
   compensate for this situation.

shopt -s nocasematch was used rather than to perform casewise matches on
the distro string for every possible upper/lower case mix.