getbit.cpp: refactored into class
Takes a decimal bit number and a hexadecimal mask without the
leading 0x and returns a 0 if the bit number is set in the mask,
or returns a zero if the bit number is not set.

Also returns a zero if the inputs are not valid decimal and hex
numbers, respectively.

Example:

getbit 10 ffff400
echo $?
1

getbit 10 ffff4000
echo $?
0