patbatcmp: more refactoring for better accuracy
Can't count on submitters to use git format-patch, so abandoning the
possibility of comparing the diffstats. AFAIK, nobody does that anyway.

Also needed a more accurate detection of end-of-patch. Because bash read
strips leading and trailing white space, IFS had to be nulled to capture
these leading and trailing whitespace to make the comparisons exact.

Also, end-of patch is a "-- " sequence, which has a trailing white
space.