gitcherrypick: Added ability to choose which hash in the commitsfile
Rather than automatically assuming that the user wants to start from the
last succesfully applied cherry-pick, give the user the option to choose
which hash in the commits file to start with.

Also fixed the problem with ambiguous and invalid hashes.

TODO:
There is still a lot of cleanup left, and the manual text needs to be
updated, but this is a pretty good picture of where it should go.