gitcherrypick: fix git status bug
git status --porcelain was returning nonzero for untracked files, which
are not the stuff of cherry-pick conflicts. The script was interpreting
any git status --porcelain content as due to git cherry-pick conflicts.
Used the returned status from git cherry-pick to make this determination
instead.

In process:
Create downstream patchset.
The idea is to start from the last downstream commit before the
gitcherrypick began, but to also allow the user to select a different
commit from which to start the backport.