gitcherrypick: major refactoring effort
Want to make the creation of the downstream patchset more intuitive
and less reliant on what goes on during the cherrypicking process.

New menu item to select the hash from which to start the call to git
format-patch.

Removed the option to start cherry-picking from last successful commit,
because it was redundant, given the option to (S)elect the commit in
the upstream commits file.

TODO:
Implement the code that actually creates the downstream patch set.