lib/patch-common: some refactoring
Move exec_patcmp out of patchreview and patchdevelop to patch-common.
Simplify the invocations of the path to the upstream commits file by
providing a static name and then updating the dynamic us_commits and
bp_commits variables with each pass through the loop, by adding the
working directory to the filename.