patcmp: much refactoring to fix bug in indexing code
Fixed the bug where selecting a specific patch number outside of the
mismatch_array when in mismatch_mode caused problems returning to the
correct spot in the mismatch_array.

Removed a lot of cruft code that wasn't gonna be used.

TODO:

Still do not land in the correct patch when pressing p to update the
upstream commit number.

Need to remove or comment-out the debug code.