lib/ui.soure: ui_inarray() remove echoes
Do not echo anything, just return 0 if the number is in the array or 1
if the number is not in the array.

The reason for this change is that the third parameter would not be
updated if used in a shell test:

	[ $(ui_inarray num array[@] index) -eq 0 ]

In such a use case, the "index" argument would not be updated.

Also if the number were not found, the third parameter would be
left unmolested.