patcmp: get the indexes working properly
Fixed a couple of bugs, one being the use of the -n and -k switches
in the call to git format-patch in replace_commit().

The other bug that was fixed was backing up from the ultimate patch
when wanting to replace the ultimate patch. Originally, when trying
to replace the ultimate patch, the penultimate patch always got
replaced. The fix was not to backup if we are on the ultimate patch,
because the index will not be incremented once we have incremented
it to its ultimate value.

The problem with that fix is that the index advances to the maxindex
after viewing the penultimate patch. If we test for index, then we will
replace the ultimate patch, not the penultimate. The previndex variable
tracks the most recently viewed, so we know we can backup to that if it
was not the ultimate patch in the set.

Also changed the logic so that if you examine a patch outside of the
mismatched set, with mismatch set true, the next patch to be examined
will be the next one that would have been examined in the mismatch
set.