gitcherrypick: redesign and refactoring continue
A lot of work remains to be done, since the way this works is being
changed dramatically.

TODO:
Must retreive all the sticky variables at the start of the session
unless running from the menu. Each cycle in the menu must update all the
sticky variables, so there needs to be a function that recalls all the
sticky variables when --continue is used.

Need to add a --commit to return to the script after resolving a
conflict. The script still needs to exit with a message when
encountering a conflict.

Need to implement the --abort switch to abort a conflicted cherrypick
and reset the script's sticky variables to their pre-conflict values.
This should involve nothing more than setting the lastuphash to the
most recent successfully applied hash and clearing the b_conflict flag.