cherp: made some changes to cherrypick and stage_commit
No need to pull from the project file, since that's done once for
all when the script starts.

Call stage_commit() from cherrypick()

There was a bug in init_session where you could invoke cherp with no
args when there was no project file. This would end up in a hang in
ui_readline, because the upcmtstartline had not been initialized.

Fix was for status in init_session from calls to routines. If bad
status, then call init_newsession to initialize the world.

If there was no project file, then init_newsession is the right
thing to do.

TODO:
Perhaps it would be best to check for the existence of a project
file from the start. If a --stat or --history call, then exit
with a notification to invoke cherp --new.

If it's not a --stat or --history call, then invoke init_newsession.

Also need to fix ui_readline.

Return with nonzero if any of the following conditions are not met.
There must be exactly two args
The first arg must be a number
The second arg must be a valid file [ -f $2 ]