cherp: added run_single and pop_commits routines
Also cleaned up some of the logic.

Need to call get_next_cherrypick upon every entry, whether init_session
or init_newsession. In init_session, I had to decrement the
upcmtstartline, and in init_newsession, the upcmtstartline has to be
inited to zero. That's because get_next_cherrypick increments the
upcmtstartline every time it's called. May have to rethink this and
use an arg to tell it whether to increment or not.

In chk_prj_file, needed to check some of the fields being read from
the project file. If these fields are empty, then init_newsession
must be called to init them.

Cleaned up run_me and removed from the project file some of the
variables that don't need to be sticky.

Added --abort, --reset, and --pop to the command line options

Started testing with --pick <commit> and got it working.