cherp: added pop_upcmtstartline()
When popping commits on the downstream branch, the line counter for
the upstream commits file must also be decremented. This isn't as
easy as it sounds, because we have to skip over lines beginning with
a #. For example, if we pop 5, and one of the lines in the upstream
commits file has a hash in it, we must decrement the line counter
by 6.

I preferred to let the line counter for the upstream commits file
count every line in it, even the hashed-out ones. In the end, it's
easier to keep track of things this way.

Fixed some of the logic of the externally called rouines and
factored common init steps out of the init_session, init_newsession,
and init_cmdline_call routines.

Also did some refactoring, moving things around, and comment repair.