cherp: fix command line to get correct optional argument
We weren't able to see the second argument when first argument didn't
have leading double-dash "--", because the first argument wasn't
consumed by option parsing loop. So if the user enters a first argument
without the double-dash, then shift to get the original first argument
out of the way so the option parsing loop can access the second arg.

TODO:
Get rid of the double-dash altogether, because these are not really
options. They are mutually exclusive commands, so parse them like
commands. It's possible that they will even have their own options in
the future.