ui.source: add loop_range_q to give quit option to caller
Need to have a separage loop_range function if we want to give the
user the option to quit without making a selection.

There are some situations where a selection absolutely must be made.
For that, there is loop_range.

For situations where the user is allowed to return without making a
selection, there is now loop_range_q