patchtools: remove booleans and add Rename and Format controls
Major refactoring effort. The booleans for renaming and formatting
patches were removed and replaced with controls. The underlying
booleans still exist, but they are automatically manipulated, as
before, but with the additional option to run the Rename and Format
functions from the interface.

Also removed the option to change the 40-char hash to 7-bit, since
the RHEL kernel maintainers now require the full 40 chars of the
upstream commit hash.

This is moving closer to the ability to implement a CLI.

lib/patch-common.source
patchdevelop
patchreview