replacelinen: replace any text, even with special chars
Problem with doing this with sed or ark is that the regex can be
terminated by typical regex delimiters or wildcards, unless they
are escaped. After a while, such code becomes indecipherable and
unmaintainable.

This utility bypasses all of that hassle, because the whole string
can be taken literally by using single quotes. No such protection
exists in utilities using regexes.