patchtools: patchcommon: more patch prefix accommodation
New patch prefix variant: [PATCH RHEL 7.6

So, now we need to deal with PATCH coming before the RHEL and version,
as well as a space as well as the possibility of a hyphen between
the RHEL and version.

Permutations and combinations of all the possible ways to juggle
these elements, including hyphens, and/or spaces, is in the hundreds.

Hoping that this fix catches them all.