newbashscript: all new scripts will at least have -h option
This creates a script template with at least a -h "help" option, as
well as minimal help text based on the options and args passed to
this script template generator.

The script is also modular, with a call to a main function, which
encourages using functions for everything.

There are some global exit codes defined, as well as Bold and
Underline terminal codes, as well as a default exit function and
control-c detection.