waitsecs: initial entry
Wait a number of seconds specified by the caller. The interval at
which dots are printed to the screen depends on the number of
seconds. Ten or less, and 10 dots per second. Between 10 and 100
seconds, the dots are printed in half-second intervals. For waits
greater than 100 seconds, dots are printed at one second intervals.