findpkgs: removed artificial search of hyphenated package variants
This was more than it needed to be. The more important thing was to
find the main package, rather than any of its hyphenated variants.
The restriction to exact package name was removed, so all the
hyphenated variants could be discovered, but it wasn't necessary to
look for them in particular, since their presence indicates support
for the main package in the given stream.

Left some debug code in place, but commented out.