patrmhead: major refactoring
This facility never really worked well. Now we have it working.
One of the problems was the archive removes the @ and . from
email addresses to prevent farming of those addresses by web bots.
Malformed addresses crater 'git am', so we added code to put the
@ and . back where we think they should be. All that's important
with this utility is to make 'git am' happy.