patchtools: gitutilities: git_comment2commitsfile: smarter commit parsing
We were getting a lot of false positives on what appeared to be multiple
lines containing 40-char hashes.

Problem was solved by only considering first those lines beginning with
the token '(cherry'. If there were none of those, then look for lines
beginning with the token 'commit'.

The squashing of duplicate commits was left unchanged.