lib/patch-common.source: skip prologues and looser patch validation
Patch validation consisted of looking for start maker 'diff --git'.
Unfortunately, not all patches are formatted this way, so that
validation was removed.

There was an unused function, check_for_prologues. That function
was renamed to is_prologue and modified to return a 0 if the patch
is the zeroth patch in a set, and 1 if it is not.

Function create_patchlist() now makes the call, but for only the first
patch in the list.

If there is no slash in the Subject line, then it is not a multi-patch
series, so the function returns 1.