patchtools: more robust extract_subject routine
There are some patch submissions that have subject lines that are
formatted:
[something][something else]

Instead of looking for the first ']', now look for the very last ']'.
This could bite in the future, if someone submits a subject line with
right square bracket in the subject proper.

Added ui_getlastch() to ui_source.

Fixed bug in patch-common::is_prologue where a / could be located
between two strings, instead of two numbers, as in "power/pc"

TODO:
   Make this more robust by looking only between the brackets for
   the patch number rather than the whole subject string.