patchtools: more commit validation work
In git_utilities extract_subject() and git_comment2commitsfile()
The subject line can sometimes have embedded cruft in it from the
mail server or whatever the submitter uses to create his patches.
Truncate the subject line to 80 characters as a stop gap measure.

TODO:
   See if we can find one congruent paragraph in the upstream commit
   log vis-a-vis the submitted RHEL patch. This will slow things a
   bit, but it's better than lists of commits to choose from on one
   hand and failure to validate a commit on the other.

In patch-common format_upstream_patches()
Give the user the option to reformat the upstream patches by deleting
the contents of the directory containing them. The user is presented
with a query of whether he wants to delete the contents of the
directory.

In ui loop_yn()
This never worked as intended, so fixed it.

	modified:   lib/gitutilities.source
	modified:   lib/patch-common.source
	modified:   lib/ui.source