patchtools: validate commit fixes
Was relying on the last ] in the subject line to determine where to
start tokenizing the line. Unfortunately, upstream committers use
[] in their subject lines, too.

Instead, tokenize the whole RHEL subject line and look for the first
occurance of the second token in the upstream subject line. If none
is found, then the commit is not valid.

If a match is found, see if the rest of the upstream subject line
matches word for word with that much of the RHEL subject line,
starting with the first token match.

If they are not the same, then the commit is invalid.