patchtools: use the tokens in the RH patch subject line
Encountered a case where the RH subject line was truncated. Because we
were using the number of upstream tokens as the max tokens in the
comparison loop, we got a mismatch when we exhaused the tokens in the RH
token array.

Change the max tokens to the count in the RH token array, so we can get
a match even when the RH subject line has been truncated.

This increases the likelihood of redundancies, but those will be
presented to the user to sort out. Better than not having a match at
all.

	modified:   lib/gitutilities.source