patchtools: refactoring and better array management
Use the functions in patch-utils rather than having copies of them
in more than one place. Use of of a valid patch list and its array
is the best way to deal with prologue patches. Made this universally
available in patch-utils with create_patchlist.

Check whether a source script has already been loaded.