patchtools: added fuzz option to patchreview commit validation
Patches are validated in two steps.
First, upstream commit hashes in the submitted RHEL patch are looked up.
Second, the summary line from 'git log --oneline' of the upstream hash is
compared with the summary line of the submitted RHEL patch. This is done
to avoid confounding the mention of any hash in the git log with the one
being backported.

Because there can be some minor differences in the way backports copy the
upstream summary line, fuzz is sometimes necessary to minimize false
positives of invalid commits.

Made some enhancements to other parts of the patchreview tool to better
explain output, etc.

lib/gitutilities.source
lib/patch-common.source
lib/ui.source
patchreview