lib/gitutilities.source: revisit commit validation logic
Upstream commits are now gleanded from each patch submitted as follows.

For each patch:

Gather anything that looks like a commit into a scratchfile

See if there is a commit we can successfully validate to
minimize user pain.

If none of the commits can be validated, then gather what
there are and present them to the user for selection.

If there are no commits in the patchfile's commit log, then
we won't be able to format an upstream commit corresponding
to this patch.