patchtools: WIP: patchreview/develop
This is a work in progress.

The problem with find_fixes as implemented was that it would never
fail, because the linus tree was included in the search.

In order to exclude the linus tree from the search, searching only
the current branch from where it first forked from the linus tree,
some significant infrastructure had to be added.

Consequently, there was some refactoring and minor bug fixing here
and there.

modified:   lib/gitutilities.source
modified:   lib/patch-common.source
modified:   lib/patch-utils.source
modified:   patchdevelop
modified:   patchreview