patchtools: patch-common: some bug fixes and improvements
. show_mrstatus
  Use the $STA color to display the MR

. call_lab_mr_list
  Use 'lab mr list' instead of 'lab mr list --all'. We don't want
  to see a bunch of inactive MRs.

. select_merge_request
  Improve messaging.

. get_merge_request_patches
  Don't bother saving files in the indir for MR mode.
  Always write new ones.

. format_upstream_patches
  Always write new files into the outdir in MR mode.

One of the reasons for always creating new files in MR mode, is that
it would take extra logic to process the MR patches without writing
them into the indir. Consequently, it seemed better to apply the
same approach to the outdir files. That is, just create them from
scratch every time.