patchtools: add filter MR by author
patchreview: modify the 'g' grep menu item to filter for either the
author of an MR or for a string in the MR subject line.

lib/patch-common.source: add support for the new functionality. Also
run a patch download and compare if a selection is made from the
filtered list.

lib/ui.source: modify the qstr to have a more consistent appearance.