patchtools: allow n/ack comments on any MR
The n/ack, comments option 'a' is now always available from the
main menu. However, it will always ask for the specific MR, rather
than assume the most recent one reviewed. Reviewed MRs automatically
invoke the n/ack/comments menu after having been reviewed.

Default for new installs is not to show the Discrete Operations
display. Of course, this can be toggled on and off and is sticky
between invocations of the script.