patchtools: patch-mrutilities: use git instead of cki for patchcount
Using the cki bot to obtain the patchcount is not as reliable as
using git to do it. Problem with using the cki count is that MRs
are dynamic and subject to additions, deletions, and modifications.
The most recent cki count only reports the latest additions, not
the count of the series as a whole.