patchtools: patch-mrutilitie: failover patchcount
First try to extract patchcount from the CKI analyzed line. If nothing
there, then use the git merge-base method.

The reason for this is that merg-base for earlier MRs is much farther
back than where the earlier MRs were spun. This is a better way to
obtain the patchcount, because the main branch does not have to be
updated in order to get the latest MRs.

But ARK does not use CKI commit analysis, so we must rely on the
git merg-base method when there is no CKI commit analysis.

Also fixed a minor bug with the get_merge_request. If the status
value is greater than the "norev" value, then it's just a count
of remaining comment lines, so return 0.