patchtools: patch-mrutilities: count the actual patches
patchcount global has been determined by cki bot count, by finding
the merge-base between the MR and the main, and they have both
failed to give consistent results.

Actually counting the patches, using 'lab mr show -p' to parse the
actual patches for a count seems to be the most consistent so far.
The method depends on finding a redhat.com author at the beginning
of a line to detect the beginning of a patch.