patchcount.py: python script that counts patches in an MR
Initial entry.

The intent is to adapt this script to be used with patchreview in
order to extract the correct number of patches in an MR without
resorting to git commit offset and merg-base arithmetic. Those do
not present consistent results and have caused some problems in
the past, especially with 'git merge-base'.