patchtools: patch-common: menu spacing and verbose changes
Adapt menu spacing for a menu line that needed an extra space.

Adapt to the new options interface for git_comment2commitsfile.
This was implemented to allow for more than one optional arg to
the func in gitutilities. The optional arg was needed to accomodate
verbose/quiet output options.

modified:   lib/patch-common.source